Remove usage of GIT_MERGE in examples
The GIT_MERGE environment variable is of no use. It might be that it was
supposed to be GIT_MERGE_MASTER_BRANCH, but it makes no sense to merge
the master branch in master, if both master branches are from the same
git URL... So, just drop this from the examples.