Check current branch
For some reason I'm not too sure about, seveal projects have their
default branch as 'el6'.  Trying to "git checkout -t -b foo
origin/foo" when you're already on foo fails.

This adds a sanity check to avoid switching branches if already on the
right one.  This makes things more robust should upstream change their
default branch; this allows us to still specify the branch to choose
in the build part

This also removes the check that skips checkout if already on master
branch.  For those branches where we are checkout on el6 by default,
we want to switch back to master.