Create .gitconfig in Xen dom0 when merging.
We recently discovered that git master merging was failing in the dom0.
The issue was due to the fact that no git email address was set and
thus the merge would fail.

This would cause some of the upstream reviews to fail to install
plugins if they posted new code but didn't rebase it (thus causing a
merge)